!208 清理版本级 cleancode * 检视意见修改 * 检视意见修改 * cleancode去除 * 去除无效注释 * 清理版本级 cleancode * 清理版本级 cleancode

+++ src/smem/csrc/common/smem_monotonic_time.h
@@ -112,14 +112,14 @@ inline uint64_t MonotonicTime::TimeUs()
 {
     struct timespec ts{};
     clock_gettime(CLOCK_MONOTONIC, &ts);
-    return static_cast<uint64_t>(ts.tv_sec) * 1000000L + ts.tv_nsec / 1000L;
+    return static_cast<uint64_t>(ts.tv_sec) * 1000000L + static_cast<uint64_t>(ts.tv_nsec) / 1000L;
 }
 
 inline uint64_t MonotonicTime::TimeNs()
 {
     struct timespec ts{};
     clock_gettime(CLOCK_MONOTONIC, &ts);
-    return static_cast<uint64_t>(ts.tv_sec) * 1000000000L + ts.tv_nsec;
+    return static_cast<uint64_t>(ts.tv_sec) * 1000000000L + static_cast<uint64_t>(ts.tv_nsec);
 }
 #endif /* ENABLE_CPU_MONOTONIC */
